+++ views/amendment/_set_proposed_procedure.php
@@ -23,25 +23,10 @@ echo Html::beginForm($saveUrl, 'POST', [
     'data-context'             => $context,
     'data-proposal-id'         => ($proposal->isNewRecord ? null : $proposal->id),
 ]);
-if ($proposal->proposalStatus === Amendment::STATUS_REFERRED) {
-    $preReferredTo = $proposal->comment;
-} else {
-    $preReferredTo = '';
-}
-if (in_array($proposal->proposalStatus, [Amendment::STATUS_OBSOLETED_BY_AMENDMENT, Motion::STATUS_OBSOLETED_BY_MOTION])) {
-    $preObsoletedBy = $proposal->comment;
-} else {
-    $preObsoletedBy = '';
-}
 if ($proposal->proposalStatus === Amendment::STATUS_PROPOSED_MOVE_TO_OTHER_MOTION) {
     $preMovedToMotion = $proposal->comment;
 } else {
     $preMovedToMotion = '';
-}
-if ($proposal->proposalStatus === Amendment::STATUS_CUSTOM_STRING) {
-    $preCustomStr = $proposal->comment;
-} else {
-    $preCustomStr = '';
 }
 
 if (isset($msgAlert)) {
@@ -111,31 +96,13 @@ $limitedDisabled = ($canBeChangedUnlimitedly ? null : true);
 
     <?= $this->render('../shared/_proposed_procedure_tags', ['imotion' => $amendment]) ?>
 
-    <section class="statusDetails status_<?= Amendment::STATUS_OBSOLETED_BY_AMENDMENT ?>">
-        <label class="headingLabel"><?= Yii::t('amend', 'proposal_obsoleted_by') ?>...</label>
-        <?php
-        $options = ['-'];
-        $filter = IMotionStatusFilter::onlyUserVisible($consultation, false);
-        foreach ($amendment->getMyMotion()->getVisibleAmendmentsSorted() as $otherAmend) {
-            if ($otherAmend->id !== $amendment->id) {
-                $options[$otherAmend->id] = $otherAmend->getTitle();
-            }
-        }
-        foreach ($filter->getFilteredConsultationIMotionsSorted() as $otherMotion) {
-            if ($otherMotion->id === $amendment->motionId) {
-                continue;
-            }
-            if (!is_a($otherMotion, Motion::class)) {
-                continue;
-            }
-            foreach ($otherMotion->getVisibleAmendmentsSorted() as $otherAmend) {
-                $options[$otherAmend->id] = $otherAmend->getTitle();
-            }
-        }
-        $attrs = ['id' => 'obsoletedByAmendment', 'disabled' => $limitedDisabled];
-        echo Html::dropDownList('obsoletedByAmendment', $preObsoletedBy, $options, $attrs);
-        ?>
-    </section>
+    <?= $this->render('../shared/_proposed_procedure_status_details', [
+        'imotion' => $amendment,
+        'proposal' => $proposal,
+        'limitedDisabled' => $limitedDisabled,
+        'canBeChangedUnlimitedly' => $canBeChangedUnlimitedly,
+    ]) ?>
+
     <section class="statusDetails status_<?= Amendment::STATUS_PROPOSED_MOVE_TO_OTHER_MOTION ?>">
         <label class="headingLabel"><?= Yii::t('amend', 'proposal_moved_to_other_motion') ?>:</label>
         <?php
@@ -156,36 +123,7 @@ $limitedDisabled = ($canBeChangedUnlimitedly ? null : true);
         echo '<div>' . Yii::t('amend', 'proposal_moved_to_other_motion_h') . '</div>';
         ?>
     </section>
-    <section class="statusDetails status_<?= Amendment::STATUS_REFERRED ?>">
-        <label class="headingLabel" for="referredTo"><?= Yii::t('amend', 'proposal_refer_to') ?>...</label>
-        <input type="text" name="referredTo" id="referredTo" value="<?= Html::encode($preReferredTo) ?>"
-            <?php if (!$canBeChangedUnlimitedly) echo 'disabled'; ?>
-               class="form-control">
-    </section>
-    <section class="statusDetails status_<?= Amendment::STATUS_CUSTOM_STRING ?>">
-        <label class="headingLabel" for="statusCustomStr"><?= Yii::t('amend', 'proposal_custom_str') ?>:</label>
-        <input type="text" name="statusCustomStr" id="statusCustomStr" value="<?= Html::encode($preCustomStr) ?>"
-            <?php if (!$canBeChangedUnlimitedly) echo 'disabled'; ?>
-               class="form-control">
-    </section>
-    <section class="statusDetails status_<?= Amendment::STATUS_VOTE ?>">
-        <div class="votingStatus">
-            <h3><?= Yii::t('amend', 'proposal_voting_status') ?></h3>
-            <?php
-            foreach ($consultation->getStatuses()->getVotingStatuses() as $statusId => $statusName) {
-                ?>
-                <label>
-                    <input type="radio" name="votingStatus" value="<?= $statusId ?>" <?php
-                    if ($amendment->votingStatus == $statusId) {
-                        echo 'checked';
-                    }
-                    ?>> <?= Html::encode($statusName) ?>
-                </label><br>
-                <?php
-            }
-            ?>
-        </div>
-    </section>
+
     <section class="publicExplanation">
         <h3><?= Yii::t('amend', 'proposal_public_expl_title') ?></h3>
         <?php